# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "API REST de EPAYCO2",
    "version": "1.0.0",
    "title": "EPAYCO2 REST API",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "dmy.leomar@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "schemes": ["http", "https"],
  "paths": {
...
    "/account/recharge": {
      "post": {
        "tags": ["account"],
        "summary": "recharge account",
        "operationId": "recharge",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Recharge"
            }
          }
        ]
      }
    },
...
  },
  "definitions": {
...
    "Recharge": {
      "type": "Request",
      "required": ["document", "phone", "amount"],
      "properties": {
        "document": {
          "type": "string",
          "description": "User document"
        },
        "phone": {
          "type": "string",
          "description": "User phone"
        },
        "amount": {
          "type": "number",
          "description": "amount"
        }
      },
      "xml": {
        "name": "Recharge"
      }
    },
...
  }
}